4.0.0 Defered Regl components
This commit updates all dependencies and rewrites the library to
support direct rendering of regl "drawCommands" or components. This
update incorporates the `defered-regl` utility to enable an experience
in which developers write regl code as they would in a regular regl
project and can then pass it to a react tree to render.